chore: Adds Shadcn Components and builds out further logic
Adds a Tauri-side Rust library to scan if running in KDE to read the KDE color configuration and apply those to the CSS. This was just a curiosity and flex.

Added a few ShadCN components that I know I'm going to need.

Builds out more of the "pages" of the app.