chore: Build out Rust utilities more
Needed to restructure the ssh config parser library some more. The file load and parsing logic now builds a reasonable tree of standardized nodes that capture all data from the supplied config. This include Host, Include and Comment stanzas. We also capture any "invalid" parameters in the tree to make sure we don't lose information. The trick was trying to think like the ssh client and read the file backwards since that's more or less how ssh treats the file. More "global" settings are at the end of the config file.