chore: Render loaded ssh config files
Now that the load file functions in the Rust side of things, time to load that tree into the UI. Got a start on it. It's far from pretty and there is still some way to go with it. But at least I've mostly figured out the components I will need. Just need to rearrange things a bit.